replaced CoordinatorLayout in MainActivity with MainFragment class
