+++ app/src/main/res/values/dimens.xml
@@ -6,5 +6,6 @@
     <dimen name="palette_offset_x">6dp</dimen>
     <dimen name="palette_offset_y">40dp</dimen>
 
+    <dimen name="searchfilter_button_ripple">22dp</dimen>
     <dimen name="theme_button_ripple">26dp</dimen>
 </resources>
